// src/features/project/interfaces.ts
export enum IProjectTubsName {
	All_ABOUT_PROJECT = 'project.project_all_details',
	LAYAOUT = 'project.layout.title',
	COMPOSITION_OF_PROJECT = 'project.composition.title',
	ALTERNATIVE = 'catalogue.changes.title',
	ADDITIONAL_SERVICES = 'header.additional_link',
	PROJECT_IN_PROGRESS = 'project.project_in_progress',
	SIMILAR_PROJECTS = 'about_us.similar_projects.title'
}

export const tubsArray: IProjectTubsName[] = [
	IProjectTubsName.All_ABOUT_PROJECT,
	IProjectTubsName.LAYAOUT,
	IProjectTubsName.COMPOSITION_OF_PROJECT,
	IProjectTubsName.ALTERNATIVE,
	IProjectTubsName.ADDITIONAL_SERVICES,
	IProjectTubsName.PROJECT_IN_PROGRESS
]